+++ Cargo.toml
@@ -31,10 +31,9 @@ rustversion = "1.0"
 walkdir = "2.3"
 serde = "1.0"
 serde_derive = "1.0"
-serde_json = "1.0"
+serde_test = "1.0"
 
 [features]
-default = []
 example_generated = []
 rustc-dep-of-std = ["core", "compiler_builtins"]
