+++ tests/compile-pass/empty.rs
@@ -1,0 +1,10 @@
+#![no_std]
+
+use bitflags::bitflags;
+
+bitflags! { }
+
+bitflags! {
+    struct Flags: u32 {
+    }
+}
