+++ tests/compile-pass/trait/convert.rs
