+++ tests/compile-pass/trait/debug.rs
