+++ tests/compile-pass/trait/default.rs
