+++ tests/compile-pass/trait/eq.rs
