+++ tests/compile-pass/trait/inherent_methods.rs
